requirements.txt
asyncio==3.4.3
openai==1.51.1
panel==1.5.2
python-dotenv==0.19.2

